refactor(c2pa): extract shared exclusion parsing and uuid helpers
Move parseExclusions to bmff/, and toUint8Array, asInteger, matchesUuid,
JUMBF_UUID to utils so upcoming Merkle validation can reuse them.
